moves requests form /components folder to /sagas folder and renames all-data to both start with getGroups
